# docs/swagger.yaml
definitions:
  models.ChangePasswordRequest:
    properties:
      new_password:
        type: string
      old_password:
        type: string
    type: object
  models.Comment:
    properties:
      content:
        type: string
      review_id:
        type: integer
      user_id:
        type: integer
    type: object
  models.Feature:
    properties:
      details:
        type: string
      name:
        type: string
    type: object
  models.LoginRequest:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
...
  models.RegistRequest:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
...
info:
  contact: {}
paths:
  /auth/change-password/{id}:
    put:
      consumes:
      - application/json
      description: Change user password
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: Password
        in: body
        name: password
        required: true
        schema:
          $ref: '#/definitions/models.ChangePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Change user password
      tags:
      - auth
  /auth/login:
    post:
      consumes:
      - application/json
      description: Log in a user
      parameters:
      - description: Login
        in: body
        name: login
        required: true
        schema:
          $ref: '#/definitions/models.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Log in a user
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Register a new user
      parameters:
      - description: Regist
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.RegistRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Register a new user
      tags:
      - auth
  /comments:
    post:
      consumes:
      - application/json
      description: Create a new comment
      parameters:
      - description: JWT Authorization header
        in: header
        name: Authorization
        required: true
        type: string
      - description: Comment
        in: body
        name: comment
        required: true
        schema:
          $ref: '#/definitions/models.Comment'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Comment'
      security:
      - ApiKeyAuth: []
      summary: Create a new comment
      tags:
      - comments
  /comments/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a comment
      parameters:
      - description: JWT Authorization header
        in: header
        name: Authorization
        required: true
        type: string
      - description: Comment ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Delete a comment
      tags:
      - comments
    put:
      consumes:
      - application/json
      description: Update a comment
      parameters:
      - description: JWT Authorization header
        in: header
        name: Authorization
        required: true
        type: string
      - description: JWT Authorization header
        in: header
        name: Authorization
        required: true
        type: string
      - description: Comment ID
        in: path
        name: id
        required: true
        type: integer
      - description: Comment
        in: body
        name: comment
        required: true
        schema:
          $ref: '#/definitions/models.Comment'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      - ApiKeyAuth: []
      summary: Update a comment
      tags:
      - comments
  /comments/{review_id}:
    get:
      consumes:
      - application/json
      description: Get comments by review ID
      parameters:
      - description: JWT Authorization header
        in: header
        name: Authorization
        required: true
        type: string
      - description: Review ID
        in: path
        name: review_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Comment'
            type: array
      security:
      - ApiKeyAuth: []
      summary: Get comments by review ID
      tags:
      - comments
...
  /phones/{phone_id}/features:
    post:
      consumes:
      - application/json
      description: Create a new feature
      parameters:
      - description: JWT Authorization header
        in: header
        name: Authorization
        required: true
        type: string
      - description: Phone ID
        in: path
        name: phone_id
        required: true
        type: integer
      - description: Feature
        in: body
        name: feature
        required: true
        schema:
          $ref: '#/definitions/models.Feature'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Feature'
      security:
      - ApiKeyAuth: []
      summary: Create a new feature
      tags:
      - features
  /phones/{phone_id}/features/{feature_id}:
    delete:
      consumes:
      - application/json
      description: Delete a feature of a phone
      parameters:
      - description: JWT Authorization header
        in: header
        name: Authorization
        required: true
        type: string
      - description: Phone ID
        in: path
        name: phone_id
        required: true
        type: integer
      - description: Feature ID
        in: path
        name: feature_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Feature'
      security:
      - ApiKeyAuth: []
      summary: Delete a feature of a phone
      tags:
      - features
    put:
      consumes:
      - application/json
      description: Update a feature of a phone
      parameters:
      - description: JWT Authorization header
        in: header
        name: Authorization
        required: true
        type: string
      - description: Phone ID
        in: path
        name: phone_id
        required: true
        type: integer
      - description: Feature ID
        in: path
        name: feature_id
        required: true
        type: integer
      - description: Feature
        in: body
        name: feature
        required: true
        schema:
          $ref: '#/definitions/models.Feature'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Feature'
      security:
      - ApiKeyAuth: []
      summary: Update a feature of a phone
      tags:
      - features
...
swagger: "2.0"